更新TileMap
将原先TileMap拆分成TileMap、TileMapRenderer

+++ NoaEngine/NoaEngine/TileMap.h
@@ -15,7 +15,7 @@ namespace noa {
 	public:
 		static TileMap* Create(noa::Scene* scene);
 		void AddTileMapRenderer(noa::TileMapRenderer* layerRenderer);
-		bool CheckCollision(float x, float y);
+		bool CheckCollision(int x, int y);
 	private:
 		ACTOR(TileMap)
 	private:
